move validate_file to core.io
